refactor: split admin controller into item, monthlyPlant, and updateNote modules
